pyproject.toml
[tool.poetry]
name = "desafio-lacrei-saude"
version = "0.1.0"
description = ""
authors = ["Tiago Hinterholz <fh.tiago@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.1.1"
django-cors-headers = "^4.9.0"
djangorestframework-simplejwt = "^5.5.1"
drf-spectacular = "^0.28.0"
gunicorn = "^23.0.0"
drf-spectacular-sidecar = "^2025.10.1"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^6.1.0"
flake8 = "^7.3.0"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
autopep8 = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
